# .history/docs/design/cash-flow-api-specification_20250905215824.yaml
openapi: 3.0.3
info:
  title: Cash Flow Management Service API
  description: |
    Pure calculation service API for cash flows in synthetic swap contracts.
    Focuses on calculation performance with zero storage overhead for input data.
    Supports real-time processing, incremental updates, and historical recalculation.
  version: 1.0.0
  contact:
    name: Core Lifecycle Management Team
    email: lifecycle-team@financial-firm.com
  license:
    name: Proprietary
    url: https://financial-firm.com/license
servers:
  - url: https://cashflow-service.internal/api/v1
    description: Internal production server
  - url: https://cashflow-service-staging.internal/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
  /cashflows/calculate:
    post:
      tags:
        - Calculations
      summary: Calculate cash flows (auto-determined type)
      description: |
        Main endpoint for cash flow calculations. The service automatically determines the calculation type based on:
        - Date range size (historical vs real-time)
        - Market data availability (incremental vs full)
        - Request context and performance requirements
        
        Calculation Type Logic:
        - REAL_TIME_PROCESSING: Single day or current date calculations (< 24 hours)
        - INCREMENTAL_UPDATE: Small date ranges (1-7 days) with existing market data
        - HISTORICAL_RECALCULATION: Large date ranges (> 7 days) or full recalculation
        
        Pure calculation service with zero storage overhead.
      operationId: calculateCashFlows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
            examples:
              realTimeExample:
                summary: Real-time calculation example (auto-determined)
                value:
                  cashFlowRequest:
                    requestId: "CF_REQ_RT_001"
                    contractId: "SWAP_IBM_001"
                    dateRange:
                      fromDate: "2024-01-15"
                      toDate: "2024-01-15"
                      calculationFrequency: "DAILY"
                    contracts:
                      - contractId: "SWAP_IBM_001"
                        underlying: "IBM"
                        notionalAmount: 400000
                        currency: "USD"
                        equityLeg:
                          underlying: "IBM"
                          quantity: 1000
                          dividendTreatment: "REINVEST"
                        interestLeg:
                          rateType: "FLOATING"
                          index: "LIBOR_3M"
                          spread: 0.0025
                          resetFrequency: "QUARTERLY"
                    positions:
                      - positionId: "POS_IBM_001"
                        contractId: "SWAP_IBM_001"
                        totalQuantity: 1000
                        status: "ACTIVE"
                    lots:
                      - lotId: "LOT_001"
                        contractId: "SWAP_IBM_001"
                        quantity: 1000
                        costPrice: 120.00
                        costDate: "2019-01-01"
                        status: "ACTIVE"
                    paymentSchedules:
                      - scheduleId: "SCH_001"
                        contractId: "SWAP_IBM_001"
                        scheduleType: "INTEREST_PAYMENT"
                        scheduledDate: "2024-01-15"
                        leg: "INTEREST"
                        notionalAmount: 400000
                    marketDataStrategy:
                      mode: "HYBRID"
                      realTimeEndpoints:
                        enabled: true
                        timeoutSeconds: 10
                      cacheStrategy:
                        enabled: true
                        ttlHours: 24
                    marketData:
                      mode: "SELF_CONTAINED"
                      data:
                        securities:
                          - symbol: "IBM"
                            basePrice: 125.50
                            baseDate: "2024-01-15"
                            changes:
                              - date: "2024-01-15"
                                price: 125.50
                                change: 0.75
                        rates:
                          - index: "LIBOR_3M"
                            baseRate: 0.0525
                            baseDate: "2024-01-15"
                            changes:
                              - date: "2024-01-15"
                                rate: 0.0525
                                change: 0.0010
                        dividends:
                          - symbol: "IBM"
                            dividends:
                              - exDate: "2024-03-15"
                                paymentDate: "2024-03-30"
                                amount: 1.65
                                currency: "USD"
      responses:
        '200':
          description: Cash flow calculation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashFlowResponse'
              examples:
                successExample:
                  summary: Successful calculation response
                  value:
                    requestId: "CF_REQ_RT_001"
                    calculationDate: "2024-01-15"
                    calculationType: "REAL_TIME_PROCESSING"
                    status: "COMPLETED"
                    summary:
                      totalContracts: 1
                      totalCashFlows: 2
                      totalAmount: 1250.75
                      currency: "USD"
                    contractResults:
                      - contractId: "SWAP_IBM_001"
                        underlying: "IBM"
                        totalPnl: 550.25
                        totalInterest: 525.50
                        totalDividends: 175.00
                        totalCashFlows: 1250.75
                        cashFlows:
                          - cashFlowId: "CF_001"
                            calculationDate: "2024-01-15"
                            cashFlowType: "INTEREST_ACCRUAL"
                            equityLegAmount: 0.00
                            interestLegAmount: 525.50
                            totalAmount: 525.50
                            currency: "USD"
                            state: "REALIZED_UNSETTLED"
                          - cashFlowId: "CF_002"
                            calculationDate: "2024-01-15"
                            cashFlowType: "EQUITY_PNL"
                            equityLegAmount: 550.25
                            interestLegAmount: 0.00
                            totalAmount: 550.25
                            currency: "USD"
                            state: "REALIZED_UNSETTLED"
                    metadata:
                      calculationVersion: "1.0"
                      calculationEngine: "PureCalculationEngine"
                      processingTimeMs: 280
                      memoryUsageMB: 512
                      dataSource: "HYBRID"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cashflows/calculate/real-time:
    post:
      tags:
        - Real-time
      summary: Real-time cash flow calculation
      description: |
        Optimized endpoint for real-time cash flow calculations with sub-100ms response time.
        Uses pure calculation approach with no input data storage overhead.
      operationId: calculateRealTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
      responses:
        '200':
          description: Real-time calculation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashFlowResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
  /cashflows/calculate/historical:
    post:
      tags:
        - Historical
      summary: Historical cash flow recalculation (explicit)
      description: |
        Explicit historical cash flow recalculation endpoint.
        Forces HISTORICAL_RECALCULATION type regardless of date range.
        Returns calculation status immediately, results available via status endpoint.
        
        Use this endpoint when you specifically need historical recalculation
        even for small date ranges or when you want to override the auto-determination logic.
      operationId: calculateHistorical
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
      responses:
        '202':
          description: Historical calculation accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationStatus'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
  /cashflows/status/{requestId}:
    get:
      tags:
        - Status
      summary: Get calculation status
      description: |
        Retrieve the status of a calculation request, including progress and results.
        Supports both real-time and historical calculation status.
      operationId: getCalculationStatus
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the calculation request
      responses:
        '200':
          description: Calculation status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationStatus'
        '404':
          description: Calculation request not found
        '500':
          description: Internal server error
  /cashflows/status/natural-key:
    get:
      tags:
        - Status
      summary: Get calculation status by natural key
      description: |
        Retrieve the status of a calculation request by its natural key components.
        Natural key: (contractId, fromDate, toDate, calculationType)
        
        Note: calculationType is determined by the service based on the request parameters.
        If you need to find a specific calculation type, use the calculationType parameter.
      operationId: getCalculationStatusByNaturalKey
      parameters:
        - name: contractId
          in: query
          required: true
          schema:
            type: string
          description: Contract identifier
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Start date for calculations
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: End date for calculations
        - name: calculationType
          in: query
          required: true
          schema:
            type: string
            enum: [REAL_TIME_PROCESSING, INCREMENTAL_UPDATE, HISTORICAL_RECALCULATION]
          description: Type of calculation
      responses:
        '200':
          description: Calculation status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationStatus'
        '404':
          description: Calculation request not found
        '400':
          description: Invalid natural key parameters
        '500':
          description: Internal server error
  /cashflows/{contractId}:
    get:
      tags:
        - Cash Flows
      summary: Get cash flows by contract
      description: |
        Retrieve cash flows for a specific contract within a date range.
        Pure calculation service - no contract data stored, only cash flow results.
      operationId: getCashFlowsByContract
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
          description: Contract identifier
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Start date for cash flow retrieval
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: End date for cash flow retrieval
        - name: cashFlowType
          in: query
          required: false
          schema:
            type: string
            enum: [INTEREST_ACCRUAL, EQUITY_PNL, DIVIDEND, SETTLEMENT]
          description: Filter by cash flow type
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [REALIZED_UNSETTLED, REALIZED_SETTLED, ACCRUED]
          description: Filter by cash flow state
      responses:
        '200':
          description: Cash flows retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CashFlow'
        '400':
          description: Invalid parameters
        '404':
          description: Contract not found
        '500':
          description: Internal server error
...
  /settlements/pending:
    get:
      tags:
        - Settlements
      summary: Get pending settlements
      description: |
        Retrieve pending settlement instructions generated by cash flow calculations.
        Convenience endpoint for quick access to pending settlements.
      operationId: getPendingSettlements
      parameters:
        - name: counterparty
          in: query
          required: false
          schema:
            type: string
          description: Filter by counterparty
        - name: currency
          in: query
          required: false
          schema:
            type: string
          description: Filter by currency
        - name: settlementType
          in: query
          required: false
          schema:
            type: string
            enum: [INTEREST, EQUITY, DIVIDEND, PRINCIPAL]
          description: Filter by settlement type
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Page size for pagination
      responses:
        '200':
          description: Pending settlements retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '400':
          description: Invalid parameters
        '500':
          description: Internal server error
...
components:
  schemas:
    CashFlowRequest:
      type: object
      required:
        - requestId
        - dateRange
        - contracts
        - marketDataStrategy
        - marketData
      properties:
        requestId:
          type: string
          description: Unique identifier for the calculation request (synthetic key)
          example: "CF_REQ_001"
        underlying:
          type: string
          description: Top-level underlying asset symbol for market data extraction
          example: "IBM"
        index:
          type: string
          description: Top-level interest rate index for market data extraction
          example: "LIBOR_3M"
        calculationType:
          type: string
          enum: [REAL_TIME_PROCESSING, BATCH_PROCESSING, HISTORICAL_RECALCULATION]
          description: Type of calculation to perform
          example: "REAL_TIME_PROCESSING"
        dateRange:
          $ref: '#/components/schemas/DateRange'
        contracts:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
          description: Contract details for calculation
        marketDataStrategy:
          $ref: '#/components/schemas/MarketDataStrategy'
        marketData:
          $ref: '#/components/schemas/MarketDataContainer'
...
    DateRange:
      type: object
      required:
        - fromDate
        - toDate
      properties:
        fromDate:
          type: string
          format: date
          description: Start date for calculations
          example: "2024-01-01"
        toDate:
          type: string
          format: date
          description: End date for calculations
          example: "2024-01-15"
        calculationFrequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, QUARTERLY]
          description: Frequency of calculations
          example: "DAILY"
    Contract:
      type: object
      required:
        - contractId
        - underlying
        - index
        - notionalAmount
        - currency
        - type
        - startDate
        - endDate
      properties:
        contractId:
          type: string
          description: Unique contract identifier
          example: "EQ_SWAP_001"
        underlying:
          type: string
          description: Underlying asset symbol
          example: "IBM"
        index:
          type: string
          description: Interest rate index
          example: "LIBOR_3M"
        notionalAmount:
          type: number
          format: double
          description: Contract notional amount
          example: 1000000.0
        currency:
          type: string
          description: Contract currency
          example: "USD"
        type:
          type: string
          enum: [EQUITY_SWAP, INTEREST_RATE_SWAP, BOND, OPTION, FUTURE, FORWARD]
          description: Contract type
          example: "EQUITY_SWAP"
        startDate:
          type: string
          format: date
          description: Contract start date
          example: "2024-01-01"
        endDate:
          type: string
          format: date
          description: Contract end date
          example: "2024-12-31"
...
    MarketDataStrategy:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum: [HYBRID, SELF_CONTAINED, ENDPOINTS]
          description: Market data loading strategy
          example: "HYBRID"
        realTimeEndpoints:
          $ref: '#/components/schemas/RealTimeEndpoints'
        cacheStrategy:
          $ref: '#/components/schemas/CacheStrategy'
...
    MarketDataContainer:
      type: object
      required:
        - data
        - version
        - asOfDate
      properties:
        data:
          $ref: '#/components/schemas/MarketDataContent'
        version:
          type: string
          description: Market data version
          example: "1.0"
        asOfDate:
          type: string
          format: date
          description: Market data as-of date
          example: "2024-01-01"
...
    CashFlowResponse:
      type: object
      required:
        - requestId
        - calculationDate
        - calculationType
        - status
        - summary
        - contractResults
        - metadata
      properties:
        requestId:
          type: string
          description: Original request identifier
          example: "CF_REQ_001"
        calculationDate:
          type: string
          format: date
          description: Calculation date
          example: "2024-01-15"
        calculationType:
          $ref: '#/components/schemas/CalculationTypeDetermination'
          description: Calculation type determination details
        dateRange:
          $ref: '#/components/schemas/DateRange'
        status:
          type: string
          enum: [COMPLETED, PROCESSING, FAILED]
          description: Calculation status
        summary:
          $ref: '#/components/schemas/CalculationSummary'
        contractResults:
          type: array
          items:
            $ref: '#/components/schemas/ContractResult'
          description: Results for each contract
        metadata:
          $ref: '#/components/schemas/CalculationMetadata'
    CalculationSummary:
      type: object
      required:
        - totalContracts
        - totalCashFlows
        - totalAmount
        - currency
      properties:
        totalContracts:
          type: integer
          description: Total number of contracts processed
          example: 1
        totalCashFlows:
          type: integer
          description: Total number of cash flows generated
          example: 2
        totalAmount:
          type: number
          format: double
          description: Total cash flow amount
          example: 1250.75
        currency:
          type: string
          description: Summary currency
          example: "USD"
...
    CashFlow:
      type: object
      required:
        - cashFlowId
        - contractId
        - calculationDate
        - cashFlowType
        - totalAmount
        - currency
        - state
      properties:
        cashFlowId:
          type: string
          description: Unique cash flow identifier
          example: "CF_001"
        requestId:
          type: string
          description: Associated calculation request
          example: "CF_REQ_001"
        contractId:
          type: string
          description: Associated contract identifier
          example: "SWAP_IBM_001"
        calculationDate:
          type: string
          format: date
          description: Calculation date
          example: "2024-01-15"
        cashFlowType:
          type: string
          enum: [INTEREST_ACCRUAL, EQUITY_PNL, DIVIDEND, SETTLEMENT]
          description: Type of cash flow
          example: "INTEREST_ACCRUAL"
        equityLegAmount:
          type: number
          format: double
          description: Equity leg amount
          example: 0.00
        interestLegAmount:
          type: number
          format: double
          description: Interest leg amount
          example: 525.50
        totalAmount:
          type: number
          format: double
          description: Total cash flow amount
          example: 525.50
        currency:
          type: string
          description: Cash flow currency
          example: "USD"
        state:
          type: string
          enum: [REALIZED_UNSETTLED, REALIZED_SETTLED, ACCRUED]
          description: Cash flow state
          example: "REALIZED_UNSETTLED"
        equityUnrealizedPnl:
          type: number
          format: double
          description: Unrealized P&L
          example: 0.00
        equityRealizedPnl:
          type: number
          format: double
          description: Realized P&L
          example: 550.25
        equityTotalPnl:
          type: number
          format: double
          description: Total P&L
          example: 550.25
        equityDividendAmount:
          type: number
          format: double
          description: Dividend amount
          example: 175.00
        equityWithholdingTax:
          type: number
          format: double
          description: Withholding tax
          example: 17.50
        equityNetDividend:
          type: number
          format: double
          description: Net dividend
          example: 157.50
        interestAccruedAmount:
          type: number
          format: double
          description: Accrued interest
          example: 525.50
        interestRate:
          type: number
          format: double
          description: Interest rate
          example: 0.0525
        interestNotionalAmount:
          type: number
          format: double
          description: Interest notional
          example: 400000
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T10:30:00Z"
...
    CalculationStatus:
      type: object
      required:
        - requestId
        - status
      properties:
        requestId:
          type: string
          description: Calculation request identifier
          example: "CF_REQ_001"
        status:
          type: string
          enum: [PROCESSING, COMPLETED, FAILED]
          description: Calculation status
        progressPercentage:
          type: integer
          minimum: 0
          maximum: 100
          description: Progress percentage
          example: 75
        statusUrl:
          type: string
          description: URL to check status
          example: "/api/v1/cashflows/status/CF_REQ_001"
        errorMessage:
          type: string
          description: Error message if failed
        estimatedCompletionTime:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2024-01-15T11:00:00Z"
...
    SettlementResponse:
      type: object
      required:
        - settlements
        - pagination
      properties:
        settlements:
          type: array
          items:
            $ref: '#/components/schemas/SettlementInstruction'
          description: List of settlement instructions
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
          description: Pagination information
        summary:
          $ref: '#/components/schemas/SettlementSummary'
          description: Summary statistics for the results
...
    CalculationTypeDetermination:
      type: object
      required:
        - determinedType
        - reasoning
      properties:
        determinedType:
          type: string
          enum: [REAL_TIME_PROCESSING, INCREMENTAL_UPDATE, HISTORICAL_RECALCULATION]
          description: Automatically determined calculation type
          example: "HISTORICAL_RECALCULATION"
        reasoning:
          type: object
          properties:
            dateRangeDays:
              type: integer
              description: Number of days in the date range
              example: 15
            marketDataMode:
              type: string
              description: Market data loading mode
              example: "SELF_CONTAINED"
            performanceRequirement:
              type: string
              description: Performance requirement that influenced the decision
              example: "LARGE_DATE_RANGE"
            overrideReason:
              type: string
              description: Reason if calculation type was overridden
              example: "EXPLICIT_REQUEST"
        originalRequestedType:
          type: string
          enum: [REAL_TIME_PROCESSING, INCREMENTAL_UPDATE, HISTORICAL_RECALCULATION]
          description: Originally requested calculation type (if provided)
          example: "REAL_TIME_PROCESSING"
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          description: Error code
          example: "CALCULATION_ERROR"
        message:
          type: string
          description: Error message
          example: "Calculation failed due to invalid market data"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        details:
          type: object
          description: Additional error details
        requestId:
          type: string
          description: Associated request identifier
          example: "CF_REQ_001"